[refactor] drop the storage package
